+++ src/library/screencapture/ScreenCapture_Impl.cpp
@@ -149,9 +149,9 @@ void ScreenCapture_Impl::resize(int w, int h)
 
     initScreenSurface();
 
-    /* We need to close the dumping if needed. It will open a new one on next frame */
+    /* We need to let the encoder know that the screen size has changed */
     if (Global::shared_config.av_dumping) {
-        avencoder.reset(nullptr);
+        avencoder->resize(w, h);
     }
 
     LOG(LL_DEBUG, LCF_WINDOW, "Resize Screen Capture with new dimensions (%d,%d) and size %d", width, height, size);